requirements.txt
numpy<2.0.0
opencv-python==4.8.1.78
pygame==2.6.1
PyOpenGL==3.1.7
matplotlib==3.7.2
SpeechRecognition==3.10.0
pyttsx3==2.90
wikipedia==1.4.0
google-generativeai==0.3.2
pyautogui==0.9.54
psutil==5.9.6
pywin32==306
Pillow==10.0.1
requests==2.31.0
